requirements.txt
fastapi
uvicorn
requests
python-dotenv
sentence-transformers
faiss-cpu
numpy